Revert "feat: create course home api DRF (#33173)"
This reverts commit e480a79d9cbb0f653e3ea48a0f89d66c8a04bca9.
