+++ cms/djangoapps/contentstore/rest_api/v1/serializers/__init__.py
@@ -1,10 +1,8 @@
 """
 Serializers for v1 contentstore API.
 """
-from .home import CourseHomeSerializer
 from .course_details import CourseDetailsSerializer
 from .course_team import CourseTeamSerializer
-from .course_rerun import CourseRerunSerializer
 from .grading import CourseGradingModelSerializer, CourseGradingSerializer
 from .proctoring import (
     LimitedProctoredExamSettingsSerializer,
