+++ cms/djangoapps/contentstore/rest_api/v1/serializers/settings.py
@@ -4,11 +4,23 @@ API Serializers for course settings
 
 from rest_framework import serializers
 
-from .common import CourseCommonSerializer
+from openedx.core.lib.api.serializers import CourseKeyField
+
+
+class PossiblePreRequisiteCourseSerializer(serializers.Serializer):
+    """ Serializer for possible pre requisite course """
+    course_key = CourseKeyField()
+    display_name = serializers.CharField()
+    lms_link = serializers.CharField()
+    number = serializers.CharField()
+    org = serializers.CharField()
+    rerun_link = serializers.CharField()
+    run = serializers.CharField()
+    url = serializers.CharField()
 
 
 class CourseSettingsSerializer(serializers.Serializer):
-    """Serializer for course settings"""
+    """ Serializer for course settings """
     about_page_editable = serializers.BooleanField()
     can_show_certificate_available_date_field = serializers.BooleanField()
     course_display_name = serializers.CharField()
@@ -26,7 +38,7 @@ class CourseSettingsSerializer(serializers.Serializer):
     marketing_enabled = serializers.BooleanField()
     mfe_proctored_exam_settings_url = serializers.CharField(required=False, allow_null=True, allow_blank=True)
     platform_name = serializers.CharField()
-    possible_pre_requisite_courses = CourseCommonSerializer(required=False, many=True)
+    possible_pre_requisite_courses = PossiblePreRequisiteCourseSerializer(required=False, many=True)
     short_description_editable = serializers.BooleanField()
     show_min_grade_warning = serializers.BooleanField()
     sidebar_html_enabled = serializers.BooleanField()
