+++ cms/djangoapps/contentstore/rest_api/v1/urls.py
@@ -1,7 +1,8 @@
 """ Contenstore API v1 URLs. """
 
+from django.urls import path
+from django.urls import re_path
 from django.conf import settings
-from django.urls import re_path, path
 
 from openedx.core.constants import COURSE_ID_PATTERN
 
@@ -9,9 +10,7 @@ from .views import (
     CourseDetailsView,
     CourseTeamView,
     CourseGradingView,
-    CourseRerunView,
     CourseSettingsView,
-    HomePageView,
     ProctoredExamSettingsView,
     ProctoringErrorsView,
     xblock,
@@ -26,11 +25,6 @@ app_name = 'v1'
 VIDEO_ID_PATTERN = r'(?:(?P<edx_video_id>[-\w]+))'
 
 urlpatterns = [
-    path(
-        'home',
-        HomePageView.as_view(),
-        name="home"
-    ),
     re_path(
         fr'^proctored_exam_settings/{COURSE_ID_PATTERN}$',
         ProctoredExamSettingsView.as_view(),
@@ -97,10 +91,5 @@ urlpatterns = [
         'help_urls',
         HelpUrlsView.as_view(),
         name="help_urls"
-    ),
-    re_path(
-        fr'^course_rerun/{COURSE_ID_PATTERN}$',
-        CourseRerunView.as_view(),
-        name="course_rerun"
     ),
 ]
