+++ cms/djangoapps/contentstore/rest_api/v1/views/__init__.py
@@ -3,10 +3,8 @@ Views for v1 contentstore API.
 """
 from .course_details import CourseDetailsView
 from .course_team import CourseTeamView
-from .course_rerun import CourseRerunView
 from .grading import CourseGradingView
 from .proctoring import ProctoredExamSettingsView, ProctoringErrorsView
-from .home import HomePageView
 from .settings import CourseSettingsView
 from .xblock import XblockView
 from .assets import AssetsView
